Learn new language
4 points on what is going on when we study
a new language.

Use the brain mechanisms to achieve your goals
in studying new languages.

While in past I was relaying only on
books and teachers I was not aware
of what is really important to do.
How to manage it and why people usually
do it that way.

I tried to summarise some knowledge about the
topic and make some comparison with
what am I doing and what I need to change